instructions for running added
This program is designed to use 2 servers, one to encode and one to decode a text file.  If the client has a coded message then it communicates with the decoder server to get the original message.  If the client has a decoded massage then it communicates with the encoder server to get an encrypted message.